player

+++ src/main/java/com/tl/myspring/model/PlayerRepository.java
@@ -2,5 +2,5 @@ package com.tl.myspring.model;
 
 import org.springframework.data.repository.CrudRepository;
 
-	public interface PlayerRepository extends CrudRepository<Player, Long> {
+	public interface PlayerRepository extends CrudRepository<Player, Integer> {
 }
